requirements.txt
streamlit
chess==1.10.0
pillow==10.3.0